prov/bgq: Fixes for fi_getinfo to support new capability setting method in MPICH
We were not handling the null hints correctly for fi_getinfo, and when using
the default info we were not setting some of the attributes quite correctly
particularly the capabilities.  We were also getting a double-free on the
default fi_bgq_global.info structure resoved by simplifying the code in
fi_bgq_getinfo - essentially we always have just 1 provider so avoid
complexities of a list.  Also added asserts to make sure the fi_context
is big enough and we're in FABRIC_DIRECT mode.