prov/bgq: remove the 'inject' packet type
The 'inject' packet type, used exclusively for the fi_[t}inject() code
path, was barely usable after support was added for the FI_SOURCE
capability and the FI_RX_CQ_DATA mode. These features required an
additional 7 bytes of space in the packet header which had to be stolen
from the application data field for the 'inject' packet type. This
reduced the application data field from 8 bytes to 1 byte.

It is expected that a more simple code flow, cache layout, etc. as a
result of dropping the 'inject' packet type will be more beneficial
than retaining an optimization for 0-byte and 1-byte fi_[t}inject()
operations.